+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * Prevent recursive calls to dlsym (#369)
 * Fix getting monitor resolution for fake fullscreen
 * Save clock_id per conditional variable
+* sem_timedwait converts timeout relative to real time
 
 ## [1.4.0] - 2020-06-19
 ### Added